add authorization rules and cruds

+++ services/product/src/main/java/com/zel92/product/service/impl/ProductServiceImpl.java
@@ -1,22 +1,27 @@
 package com.zel92.product.service.impl;
 
 import com.zel92.product.client.InventoryClient;
+import com.zel92.product.client.UserClient;
 import com.zel92.product.dto.OrderDTO;
 import com.zel92.product.dto.request.ProductRequest;
 import com.zel92.product.dto.response.ProductResponse;
 import com.zel92.product.entity.CategoryEntity;
 import com.zel92.product.entity.ProductEntity;
+import com.zel92.product.exception.AuthorizationFailedException;
 import com.zel92.product.exception.CategoryNotFoundException;
 import com.zel92.product.exception.ProductNotFoundException;
 import com.zel92.product.repository.CategoryRepository;
 import com.zel92.product.repository.ProductRepository;
 import com.zel92.product.service.ProductService;
 import com.zel92.product.utils.ProductUtils;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -24,11 +29,14 @@ public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
     private final CategoryRepository categoryRepository;
     private final InventoryClient inventoryClient;
+    private final UserClient userClient;
 
     @Override
-    public void createProduct(ProductRequest product) {
-        ProductEntity productEntity = productRepository.save(createNewProductEntity(product));
-        inventoryClient.persist(productEntity.getProductId(), product.quantity());
+    public void createProduct(ProductRequest product, HttpServletRequest request) throws AuthorizationFailedException {
+        if (checkAuthStatus(request)){
+            ProductEntity productEntity = productRepository.save(createNewProductEntity(product));
+            inventoryClient.persist(productEntity.getProductId(), product.quantity());
+        }else throw new AuthorizationFailedException("You don't have enough permission for that");
     }
 
     @Override
@@ -38,15 +46,24 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void deleteProduct(String productId) {
-        var product = getProductEntityById(productId);
-        productRepository.delete(product);
+    public void deleteProduct(String productId, HttpServletRequest request) throws AuthorizationFailedException {
+        if (checkAuthStatus(request)){
+            var product = getProductEntityById(productId);
+            productRepository.delete(product);
+        }else throw new AuthorizationFailedException("You don't have enough permission for that");
     }
     @Override
     public List<BigDecimal> getPrice(OrderDTO orderDTO) {
         List<ProductEntity> products = productRepository.findAllByProductIdInOrderByProductId(orderDTO.productIds());
         return products.stream().map(ProductEntity::getPrice).toList();
     }
+
+    @Override
+    public List<ProductResponse> findByCategory(String category) {
+        var categoryEntity = getCategoryByType(category);
+        return productRepository.findAllByCategory(categoryEntity).stream().map(ProductUtils::fromProductEntity).toList();
+    }
+
     private ProductEntity getProductEntityById(String productId) {
         return productRepository
                 .findByProductId(productId)
@@ -61,4 +78,11 @@ public class ProductServiceImpl implements ProductService {
     private CategoryEntity getCategoryByType(String type) {
         return categoryRepository.findByType(type).orElseThrow(() -> new CategoryNotFoundException("Category not found"));
     }
+
+    private Boolean checkAuthStatus(HttpServletRequest request){
+        var token = request.getHeader(HttpHeaders.AUTHORIZATION).substring(7);
+        var user = userClient.retrieveUser(token);
+        return Objects.equals(user.role(), "SUPER_ADMIN");
+    }
+
 }
